requirements.txt
sentence-transformers
requests
beautifulsoup4
tqdm
faiss-cpu
chromadb
langchain
langchain-community
numpy